Please have a look at the Statistics
I want to compare the Fat conditions for every time point. Since the Days are not equal, I need to make a list of every time point and compare the first with the second etc. But I cannot find a way to fill the list without it being overwritten every time. I feel ashamed